+++ agent/docker/entrypoint.sh
@@ -114,7 +114,7 @@ register() {
 	wait-for-it localhost:2345 -t 60
 	wait-for-it ${hostport} -t 60
 
-	temboard-agent-register \
+	temboard-agent register \
 		--host ${TEMBOARD_REGISTER_HOST-$COMPOSE_SERVICE} \
 		--port ${TEMBOARD_REGISTER_PORT-2345} \
 		--groups ${TEMBOARD_GROUPS-default} \