+++ agent/temboardagent/__main__.py
@@ -5,6 +5,7 @@ from .cli.app import app
 
 def main():
     __import__(__package__ + '.cli.serve')
+    __import__(__package__ + '.cli.register')
     __import__(__package__ + '.cli.runtask')
     return app()