+++ agent/tests/func/test_scripts.py
@@ -1,6 +1,0 @@
-from .test.temboard import exec_command
-
-
-def test_register_help():
-    rc, out, err = exec_command(["temboard-agent-register", "--help"])
-    assert 0 == rc, err