+++ agent/tests/unit/test_register_main.py
@@ -1,8 +1,0 @@
-import pytest
-
-
-def test_big_main(mocker):
-    from temboardagent.scripts.register import main
-
-    with pytest.raises(SystemExit):
-        main(argv=['--help'], environ={})