+++ CHANGELOG.md
@@ -8,6 +8,7 @@ script to proceed.
 - New database migration engine.
 - Removed parameter of `temboard --debug` CLI option.
 - Move `temboard-migratedb` as `temboard migratedb` subcommand.
+- Move `temboard-agent-register` as `temboard-agent register` subcommand.
 - New subcommands `temboard runtask` and `temboard-agent runtask` for
   debugging.